requirements.txt
altair
df-explorer
extra-streamlit-components
importlib-metadata
Markdown
markdown-it-py
markdownlit
MarkupSafe
matplotlib
numerize
numpy
oauthlib
openpyxl
pandas
pipreqs
plotly
requests
requests-oauthlib
scipy
seaborn
streamlit-card
streamlit-extras
streamlit-keyup
streamlit-option-menu
toml
camelot-py[cv]
